requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0

# PyTorch and Torchvision
# For x86 systems (Windows/Linux), install from PyPI
torch>=2.0.0
torchvision>=0.15.0

# BoxMOT (for BoxMOT trackers - optional but recommended)
boxmot>=10.0.0

